feat: #39 number 영역 함수별 테스트 스크립트 파일 분리
