+++ __test__/number/formatPhoneNumber.test.ts
@@ -1,0 +1,43 @@
+import { describe, expect, it } from 'vitest';
+import { formatPhoneNumber } from '../../src/utils/number';
+
+describe('number > formatPhoneNumber', () => {
+  describe('휴대폰번호 X', () => {
+    it('서울번호(02), 가운데가 세자리인 경우', () => {
+      const phoneNumber = '021234567';
+      expect(formatPhoneNumber(phoneNumber)).toBe('02-123-4567');
+    });
+    it('서울번호(02), 가운데가 네자리인 경우', () => {
+      const phoneNumber = '0212345678';
+      expect(formatPhoneNumber(phoneNumber)).toBe('02-1234-5678');
+    });
+    it('첫 단위가 세자리, 가운데가 세자리인 경우', () => {
+      const phoneNumber = '0311234567';
+      expect(formatPhoneNumber(phoneNumber)).toBe('031-123-4567');
+    });
+    it('첫 단위가 세자리, 가운데가 네자리인 경우', () => {
+      const phoneNumber = '03112345678';
+      expect(formatPhoneNumber(phoneNumber)).toBe('031-1234-5678');
+    });
+    it('첫 단위가 네자리, 가운데가 네자리인 경우', () => {
+      const phoneNumber = '050512345678';
+      expect(formatPhoneNumber(phoneNumber)).toBe('0505-1234-5678');
+    });
+  });
+  describe('휴대폰번호: 010 표준 이전', () => {
+    it('첫 번호가 011, 가운데가 세자리인 경우', () => {
+      const phoneNumber = '0111234567';
+      expect(formatPhoneNumber(phoneNumber)).toBe('011-123-4567');
+    });
+    it('첫 번호가 011, 가운데가 네자리인 경우', () => {
+      const phoneNumber = '01112345678';
+      expect(formatPhoneNumber(phoneNumber)).toBe('011-1234-5678');
+    });
+  });
+  describe('휴대폰번호: 010 표준 이후', () => {
+    it('첫 번호가 010, 가운데가 네자리인 경우', () => {
+      const phoneNumber = '01012345678';
+      expect(formatPhoneNumber(phoneNumber)).toBe('010-1234-5678');
+    });
+  });
+});
